change processing to stream street-segments
Will be easy to stream other things in the future
should we determine them to be necessary.

It's important to have the `close-channels` step
after the `xml/parse-file`, as we know we've read
the whole file in and passed it on to any channels
by that point. And we need to close channels so
the writers know they are done after the channels
are drained, and can close their files properly.

Lastly, we have a `wait-for-finish` so the main
thread doesn't exit before the channels have all
been drained.